+++ lib-ltc/src/h3/systimereader.cpp
@@ -2,7 +2,7 @@
  * @file systimereader.h
  *
  */
-/* Copyright (C) 2019-2020 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2019-2021 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -148,7 +148,7 @@ void SystimeReader::ActionSetRate(const char *pTimeCodeRate) {
 			m_nFps = nFps;
 			//
 			if (m_tMidiTimeCode.nFrames >= m_nFps) {
-				m_tMidiTimeCode.nFrames = m_nFps - 1;
+				m_tMidiTimeCode.nFrames = static_cast<uint8_t>(m_nFps - 1);
 			}
 			m_tMidiTimeCode.nType = tType;
 			//
