+++ lib-ltc/src/ltcdisplayrgbpanel.cpp
@@ -2,7 +2,7 @@
  * @file ltcdisplayrgbpanel.cpp
  */
 /*
- * Copyright (C) 2020 by Arjan van Vught mailto:info@orangepi-dmx.nl
+ * Copyright (C) 2020-2021 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -96,7 +96,7 @@ void LtcDisplayRgbPanel::Show(const char *pTimecode, struct Colours &tColours, s
 	m_LineColours[0].nBlue = tColours.nBlue;
 
 	for (uint8_t i = 0; i < 4; i++) {
-		m_pRgbPanel->TextLine(1 + i, m_Line[i], 8, m_LineColours[i].nRed, m_LineColours[i].nGreen, m_LineColours[i].nBlue);
+		m_pRgbPanel->TextLine(static_cast<uint8_t>(1 + i), m_Line[i], 8, m_LineColours[i].nRed, m_LineColours[i].nGreen, m_LineColours[i].nBlue);
 	}
 
 	m_pRgbPanel->Show();
