+++ lib-midi/src/rtpmidi.cpp
@@ -2,7 +2,7 @@
  * @file rtpmidi.cpp
  *
  */
-/* Copyright (C) 2019-2020 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2019-2021 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -130,7 +130,7 @@ midi::Types RtpMidi::GetTypeFromStatusByte(uint8_t nStatusByte) {
 }
 
 uint8_t RtpMidi::GetChannelFromStatusByte(uint8_t nStatusByte) {
-	return (nStatusByte & 0x0F) + 1;
+	return static_cast<uint8_t>((nStatusByte & 0x0F) + 1);
 }
 
 int32_t RtpMidi::DecodeMidi(uint32_t nCommandLength, uint32_t nOffset) {
